+++ server_back_end/src/main/java/com/example/demo/ScoreRepository.java
@@ -20,5 +20,5 @@ interface ScoreRepository extends CrudRepository<Score, Long>{
     void deleteByTutorialAndUid(Tutorial t, Long uid);
 
     @Query(value = "SELECT SUM(Tutorial.points) from Tutorial INNER JOIN Score ON Score.tid = Tutorial.tid where Score.uid = :uid", nativeQuery = true)
-    Long findByUid(@Param("uid") Long uid);
+    int findByUid(@Param("uid") Long uid);
 }
